adding Authentication is users and journalsentris upto 21

+++ journalApp/src/main/java/com/devnilesh/journalapp/service/JournalEntryService.java
@@ -7,6 +7,7 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.event.TransactionalEventListener;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -32,7 +33,7 @@ public class JournalEntryService {
             User user = userEntryService.findByUserName(userName);
             user.getJournalEntries().add(saved);
             //  user.setPassword(null);
-            userEntryService.saveEntry(user);
+            userEntryService.saveUser(user);
         } catch (Exception e) {
             System.out.println(e);
             throw new RuntimeException("somthing gone a wrong!");
@@ -55,12 +56,23 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName)
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName)
     {
-        User user = userEntryService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x ->x.getId().equals(id));
-        userEntryService.saveEntry(user);
-         journalEntryRepository.deleteById(id);
+        boolean removed=false;
+        try {
+            User user = userEntryService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userEntryService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }
+        catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An eror occure while deleting this entry",e);
+        }
+        return removed;
     }
 
 }
